Added update functionality. Addresses #7
ReplyCRUD.py: This adds both the ability to retrive by id and update the
reply. It was essential to do both since updating a reply depends on
being able to check its status before hand. If the reply has already
been sent then there is no purpose in updating it. Better to fail and
inform the user

TestReply.py: This adds the tests to check if replies are returned
correctly by id. It also checks if the replies are updated and validated
correctly