Added cancellations to replies. Addresses #7
ReplyCRUD.py: Cancellations are basically a simple update. The only
reason I'm using a separate method for cancelling in the data strategy
is because updating will involve updating all the fields whereas
cancellation will require a single field change all the time. The
validation for this is the same as updating a reply