Added message breaker. Addresses #5
TweetAdapter.py: This fill will be used as an adapter for performing api
related functions. It serves two purposes. Firstly, it allows for
swapping out libraries so that if tweepy doesn't work, I can pick up
something else. Or maybe tweepy breaks in some methods, I can use
another library alongside tweepy. The other purpose is so that the
library and the api initialization code doesn't get mixed up everywhere
inside the code. For now the code here is empty since for testing I'm
going to be patching the methods

MessageBreaker.py: This serves the purpose of breaking apart messages
and returning it in a list. It will be the message sender's
responsibility to send them in a sequential order as replies to each
other. The message breaker checks if a message is longer than 140
characters and if it is, it proceeds to break down the message into 140
characters (including the @ mention at the start). Each message will
have the @ mention added on and trailing spaces are stripped

TestMessageSender.py: To be used for testing sending messages. Currently
this only tests for message breaking.

TestAll.py: Added TestMessageSender to the suite