Added Message Sender. Addresses #5
DataConfig.py: This introduces a new concept of DataConfig class. The
Data Config class is responsible for managing what strategies are being
used to do persistence operations. This was a concept that was to be
introduced later but I put it in now since the message sender will be
called as an external script by the cron job. Therefore it needs to be
told what data strategy to use at the time.

MessageBreaker.py: Corrected it to check if message is less or equal to
140 characters before deciding to send message without performing any
operations to break it up

MessageSender.py: The message sender gets the reply by id, breaks it
apart, sends it using the TweetAdapter, and then changes status before
persisting it.

TweetAdapter.py: Added a stub to handle sending tweets

TestMessageSender.py: Added tests to check if messages are sent. This is
done by checking if the return values are correct and by patching
methods to make sure they were called with the correct values