Changed parameters that Tweet Adapter expects. Addresses #5
With regards to #42476e8b I realised that I need to pass the user id and
user data strategy into the TweetAdapater for any method call. Therefore
edited the MessageSender, MessageBreaker, and TestMessageSender to
reflect these changes. Discovered a couple of bugs in the process.

MessageBreaker.py: Updated method signature to require the user id and
user data strategy. Made change to TweetAdapter call within the method

MessageSender.py: Updated call to MessageBreaker. Discovered that the
sendReply method was being called with incorrect parameters. The first
reply id being passed in was the actual id of the reply stored in the
database. Not the tweet id. Fixed this. Changed replyIdList to
tweetIdList since that makes more sense.

TestMessageSender.py: Added mockUserDataStrategy everywhere. This should
be refactored now into a setUp method. Also added a bunch of asserts on
the patch methods themselves to ensure that they are being called
correctly. Wondering how these are maintainable since they override the
default behaviour of the function. So if I change the function
signature, the patch objects have NO idea about this change so the tests
will still happily pass. Should research into this. At some point I
should write at least one test that actually hits the api end point and
keep it as an optional test to run.