Completed Tweet Adapter implementation. Closes #12
This allows the tweets to actually be sent and for usernames to be
retrieved.

sampletweetconfig.ini: A sample config to copy and place in src as
tweetconfig.ini. This contains the auth tokens and dev tokens

TweetAdapter.py: Updated methods to actually use the tweepy api

TestTweetAdapter.py: Tests retrieving username and posting replies.
After posting a reply it immediately destroys it