Added creation of schedules. Addresses #8
runmessage.sh: A sample shell script for now. This will be converted
later to the actual script needed to send a reply

Scheduler.py: The scheduler itself. This introduces keeping the data
strategy configuration in a config file. I could convert the message
sender to also use the config. Apart from that, this incorporates the
code of creating a cron job after converting the time to UTC

TestScheduler.py: Tests the above scenario of creating a job according
to the specified command and in the UTC timzone

sampleconfig.ini: The sample confiuration. Copy to config.ini to make it
work