Corrected MessageSender to use update reply. Addresses #7
This was a mistake where after sending a message I was saving the reply
instead of updating. After changing the method I realised that the tests
were broken since instead of updating correctly, the ReplyCRUD was
returning that the message had already been sent. This was because I was
using a single object and after changing the status to sent INSIDE
MessageSender, the ReplyCRUD update reply method was throwing a
validation error since the mock was returning an object with the
sentStatus as "sent". Corrected all to use an array of deep copies of
their relevant replyToSend Objects.