Implemented the login flow. Finally! Closes #15
Although this doesn't complete the user creation flow, the main piece
which is authenticating the user and getting the auth tokens is
complete.

The login controller should ideally transfer the tweepy calls to the
TweetAdapter but since this is authentication specific functionality and
therefore used only ONCE through the application, it doesnt have to be
there. Will move it into the Tweet Adapter eventually.