Fixed bug in Updating replies after sending
Since replies being updated after being sent would have the scheduled
time that is before the time of updating (by a few seconds at least) the
update validation throws an error. This was fixed by adding a check to
see if the reply being updated still has the status unsent. Only in this
case should the scheduled time be validated.