Fixed inability to log in to another machine. Closes #24
TestUsers.py: Added a new test to see if users are updated when trying
to create a new user if the username already exists. This also tests
that timezone information is preserved in a case like this. To do this,
also changed what value is expected from a successful save operation.
The updated key is to inidicate that the user was an existing one.

UserCRUD.py: Altered implementation of createUser to match the required
use case

UserDataStrategy.py: Implemented the new getUserByUsername method to
satisfy the requirement of the new save process. Refactored how user
object is created from an sqlite row

TestSqlLiteUserDataStrategy.py: Added a new test for testing retrieval
of user by username. Also changed tests to stop testing the save
function when saving the user. That's tested in the the method that was
made for testing saves. I was using this all this while because I was
simply copy pasting the line self.assertEqual(UserCRUD.saveUser ... )
to actually create the user. Now I just use UserCRUD.saveUser without
any asserts surrounding it.

UserController.py: Shouldn't really be in this commit. Just cleaned up
some print statements that were polluting the test logs

TestUserController.py: Updated test to expect the new value in the
return value from UserCRUD.saveUser.