Added method to TweetAdapter to get tweet by id/url. Addresses #27
TweetAdapter.py: getSingleTweet will take in either an id or a url
string. For urls the urllib.parse module is used to parse the path out.
Not sure if splitting strings by '/' is good or if I should use regex.
Will research on this and change if necessary.

If the api returns a 404 this exception is caught and an appropriate
error message is returned