Create reply screen now has default values filled. Closes #30
UserController.py: Added a method to retrieve the user's timezone
details in a way that is easily usable by the template. Currently
returns the timezone as a dictionary with the keys 'hours' and
'minutes'.

TestUserController.py: Tests the above functionality. Tests how timezone
is returned both for UTC+ and UTC- values

views.py: createReply view is now passed both the timezone details AND
the current tie (+1 hour) as parameters

createreply.html: Uses the new params passed to it to prefill the time
to post and timezone values