Added cancel and update reply to ReplyController. Addresses #18
Both the update and cancel reply functions implement user permissions.
Need to refactor these methods since there's duplicated code. Also need
to consider moving this to the core.

Apart from this, the ReplyController update reply expects replyid to be
part of the form data. Need to add this to the template.