Fixed message breaking error. Closes #34
MessageSender.py: Cleaned up some print outputs.

MessageBreaker.py: Turns out the logic written didn't really work at
all. It works only if the 140th character is a space. The test data
which was just a randomly typed string had a space on the 140th
character and therefore the logical error didn't get caught. It's fixed
with this change now. There were two changes to this. First, I was
breaking the first message at 140 characters without any checks at all.
Removed that code. Then, if the 140th character was not a space but the
141st character was, it was breaking the message too early. So changed
the breaker to check if the 141st character is a space.