Core can break messages with links. Addresses #36
MessageBreaker.py: Changed how message breaking is done when a message
has links. Using twitter-text-python, we parse the URLs out of the
message. Then we replace all the urls within the message with links that
are as long as a t.co link would be. Then the message is broken down as
usual. Once the segment to append to the messages to send list is
obtained, we replace any short urls one at a time with the original urls
that were supposed to be there. When these messages are published to
Twitter, twitter will replace them again with the actual t.co urls.

TestTweetAdapter.py: Corrected a test that was causing the SlowTests to
fail. Single tweet searches that give 404 don't return exceptions. They
return a string now saying tweet not found.