Text input for reply character count accounts for links. Closes #36
createreply.html: Added the twitter-text-js library into the page
itself. For some reason, when loading it as an external script,
regardless of ordering, there is an error and the twttr module does not
load. Need to investigate later to understand why this is happening. For
now, this module provides us with twttr.txt.getTweetLength(tweetString)
to check what the length of the tweet is when considering URL
shortening.

requirements.txt: Added twitter-text-python to requirements