Added automatic timezone selection. Closes #39
This alters the sign up flow to automatically assign the users timezone
before login is clicked. That way, we can automatically redirect the
user to the index once they have authenticated with twitter. Changes
were small but affected a lot of files.

src/__init__.py: Changed startup parameter to assign the path of static
files to /replylater/app/static. This isn't ideal since the BASEPATH now
lives in two files. Need to figure out a way to keep it in a single
location

gettimezone.js: Uses Date and getTimezoneOffset to get the user's
timezone.

login.html: Gets the users timezone, and then assigns it to hidden
inputs

updatetimezone.html: Get the timezone, and then automatically changes
the values. If there is a change, informs the user that their timezone
has automatically been selected, all they need to do is press save.
Again, not happy with this interaction entirely, but now most users
won't need to even look at this screen.

UserCRUD.py: Since we are now passing in the user's timezone during
login, we no longer take the user's existing timezone record. Instead we
override it with the value we get from the browser. Again, I'm not sure
I like this interaction. What if the user has selected a different value
from where they live? Is overriding it the right thing to do? But then
again, what if the user is just moving from one timezone to another and
they need to change it? Or what if daylight savings is coming into play
and the user has changed their system clock but has forgotten to change
the timezone?

TestUsers.py: Altered test to no longer check if timezone remains the
same when calling createUser on an existing username

UserController.py: Changed parameters to take in timezone information
and changed how the user object is constructed

TestUserController.py: Changed to reflect above

views.py:

* The timezone information from login is stored in a session
  variable till the login is complete. This is then passed into the
  UserController createUser function.
* Also changed the ending path to have a trailing forward slash on all
  the paths. Just makes more sense to do this since it covers both the
  case with and and without a trailing forward slash