Refactored message breaker. Renamed variables, split functions etc
One change that might be confusing could be the part where the urls are
passed in the format urls[:]. This is because I'm not a fan of passing
things by reference, and this is how python passes by value. Else what
happens is the urls get magically manipulated INSIDE the method where
short urls are reconverted to the original links; reading this code will
be a little confusing in the future when you are wondering whatever
happened in there.